database보안작업중

+++ app/src/main/java/com/guys_from_301/stock_game/data/GameSetDao.kt
@@ -8,7 +8,7 @@ interface GameSetDao {
     fun getAll(): MutableList<GameSet>
 
     @Query("SELECT setcash FROM GameSet")
-    fun getSetCash(): Float
+    fun getSetCash(): Int
 
     @Query("SELECT id FROM GameSet")
     fun getId(): Int
